infra/functions/budget-shutdown/requirements.txt
functions-framework==3.*
google-cloud-run==0.*
google-cloud-scheduler==2.*
